--- BaseScripts/Asset.cs
@@ -1,6 +1,0 @@
-using Godot;
-
-public partial class Asset : Resource
-{
-	public int AssetId { get; set; }
-}
