+++ Core/Asset Managements & Mods/AssetResource.cs
@@ -1,0 +1,9 @@
+namespace Cutul.Core
+{
+    using Godot;
+
+    public partial class AssetResource : Resource
+    {
+        public int AssetId { get; set; }
+    }
+}
